+++ forms-flow-idm/migration/migrate-7.2.0.py
@@ -1,10 +1,63 @@
 
 import argparse
-from utils import get_access_token, get_client_id, add_client_roles, update_client_role, role_changes_710, get_group_id, get_client_roles, assign_roles_to_group, get_groups, get_sub_groups
+from utils import get_access_token, get_client_id, add_client_roles, update_client_role, role_changes_710, get_group_id_by_path, get_client_roles, assign_roles_to_group, get_groups, get_sub_groups
 from typing import List
 
 # Roles to be added, removed, or updated for the forms-flow-web client
 roles_to_add, roles_to_remove, roles_to_update = role_changes_710()
+
+
+def _migrate_default_groups_multitenant(token, client_id, client_roles, tenant):
+    """Migrate new roles to default groups in multi-tenant mode."""
+    try:
+        # Define default groups and their new role mappings
+        default_groups_to_role_mappings = {
+            f"{tenant}-formsflow-reviewer": [
+                "manage_advance_workflows",
+                "manage_bundles",
+                "manage_templates",
+                "submission_view_history",
+                "assign_task_to_others",
+                "reviewer_view_history",
+            ],
+            f"{tenant}-approver": [
+                "assign_task_to_others",
+                "reviewer_view_history",
+            ],
+            f"{tenant}-clerk": [
+                "assign_task_to_others",
+                "reviewer_view_history",
+            ]
+        }
+
+        # Add new roles to default groups
+        for group, roles in default_groups_to_role_mappings.items():
+            group_id = get_group_id_by_path(token, group)
+            if not group_id:
+                print(f"Group {group} not found, skipping and continuing the migration.")
+                continue
+            roles_to_assign = [role for role in client_roles if role['name'] in roles]
+            assign_roles_to_group(token, group_id, client_id, roles_to_assign)
+    except Exception as e:
+        print(f"Error: {e}")
+
+def _migrate_custom_groups_multitenant(token, client_id, client_roles, tenant):
+    """Migrate new roles to custom groups in multi-tenant mode."""
+
+    groups = get_groups(token)
+    tenant_groups = [
+        group for group in groups if group["path"].startswith(f"/{tenant}")
+    ]
+    exclude_groups = [
+        f"/{tenant}-admin",
+        f"/{tenant}-formsflow-reviewer",
+        f"/{tenant}-approver",
+        f"/{tenant}-clerk",
+    ]
+    role_index = {role['name']: role for role in client_roles}
+    print("Processing custom groups...")
+    process_groups_assign_roles(tenant_groups, exclude_groups, role_index, token, client_id, tenant)
+    print(f"Custom groups migration completed for tenant: {tenant}")
 
 def _migrate_tenants(tenant_keys: List[str]):
     """Update the tenant client by adding new permissions, revising permission descriptions, and removing unused roles.
@@ -20,6 +73,14 @@ def _migrate_tenants(tenant_keys: List[str]):
             client_id = get_client_id(token, f"{tenant_key}-forms-flow-web")
             add_client_roles(token, client_id, roles_to_add)
             update_client_role(token, client_id, roles_to_update)
+            print(f"Client roles updated for {tenant_key}-forms-flow-web client.")
+
+            client_roles = get_client_roles(token, client_id)
+
+            _migrate_default_groups_multitenant(token, client_id, client_roles, tenant_key)
+            _migrate_custom_groups_multitenant(token, client_id, client_roles, tenant_key)
+            print(f"Migration completed for tenant: {tenant_key}")
+
     except Exception as e:
         print(f"Error: {e}")
 
@@ -30,27 +91,32 @@ def _migrate_default_groups(token, client_id, client_roles):
         # Define default groups and their new role mappings
         default_groups_to_role_mappings = {
             "formsflow/formsflow-client": [
-                "submission_view_history"
+                "submission_view_history",
             ],
             "formsflow/formsflow-designer": [
                 "manage_advance_workflows",
                 "manage_bundles",
-                "manage_templates"
+                "manage_templates",
             ],
             "formsflow/formsflow-reviewer": [
                 "assign_task_to_others",
                 "reviewer_view_history",
                 "manage_all_filters",
-                "analyze_metrics_view",
-                "analyze_submissions_view",
-                "analyze_submissions_view_history",
-                "analyze_process_view"
+            ],
+            "formsflow/formsflow-reviewer/approver": [
+                "assign_task_to_others",
+                "reviewer_view_history",
+                "create_filters",
+            ],
+            "formsflow/formsflow-reviewer/clerk": [
+                "assign_task_to_others",
+                "reviewer_view_history",
             ]
         }
 
         # Add new roles to default groups
         for group in default_groups_to_role_mappings.keys():
-            group_id = get_group_id(token, group)
+            group_id = get_group_id_by_path(token, group)
             if not group_id:
                 print(f"Group {group} not found, skipping and continuing the migration.")
                 continue
@@ -83,32 +149,31 @@ def add_roles_to_custom_groups(group_roles, role_index):
     if "manage_tasks" in group_roles and role_index.get("assign_task_to_others"):
         roles.append(role_index["assign_task_to_others"])
     if "view_tasks" in group_roles and role_index.get("reviewer_view_history"): 
-        roles.append(role_index["reviewer_view_history"]) # discuss and add - view_tasks also
-    # Analyze roles
-    if "view_dashboards" in group_roles:
-        roles.extend(
-            role for name in ["analyze_metrics_view", "analyze_submissions_view", "analyze_submissions_view_history", "analyze_process_view"] 
-            if (role := role_index.get(name)) is not None
-        )
+        roles.append(role_index["reviewer_view_history"])
     return roles
 
-def process_groups_assign_roles(groups, exclude_groups, role_index, token, client_id):
+def process_groups_assign_roles(groups, exclude_groups, role_index, token, client_id, tenant=None):
     """Process each custom group and assign new roles."""
     for group in groups:
         if group.get("path") in exclude_groups:
             continue  # Skip excluded groups
         # Fetch client roles specific to the group
-        group_roles = group.get("clientRoles").get("forms-flow-web")
+        if tenant:
+            # For multi-tenant, use tenant-specific client roles
+            group_roles = group.get("clientRoles").get(f"{tenant}-forms-flow-web")
+        else:
+            # For single-tenant, use the default client roles
+            group_roles = group.get("clientRoles").get("forms-flow-web")
         if not group_roles:
             continue  # Skip if no client roles
-        print(f"Group: {group.get('path')}")
+        print(f"Custom Group: {group.get('path')}")
         roles = add_roles_to_custom_groups(group_roles, role_index)
         # If new roles are found, assign them to the group
         if  roles:
             assign_roles_to_group(token, group.get("id"), client_id, roles)
         if group.get("subGroupCount", 0) > 0:
             sub_groups = get_sub_groups(token, group.get("id"))
-            process_groups_assign_roles(sub_groups, exclude_groups, role_index, token, client_id)
+            process_groups_assign_roles(sub_groups, exclude_groups, role_index, token, client_id, tenant)
     return True
 
 def _migrate_custom_groups(token, client_id, client_roles):